% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_mvrnorm.R
\name{bceMCMC_mvrnorm}
\alias{bceMCMC_mvrnorm}
\title{This is an MCMC function with a multivariate normal proposal density that is only meant to be used by the gpMCMC master function}
\usage{
bceMCMC_mvrnorm(nmcmc, burn, thin, x, y, reg, step, priortheta)
}
\arguments{
\item{nmcmc}{Number of mcmc samples to generate, before burning and thinning}

\item{burn}{number of samples to burn}

\item{thin}{keep only one of every 'thin' samples}

\item{x}{predictors}

\item{y}{response}

\item{reg}{currently only option is "constant"}

\item{step}{$(step^2)/p$ is multiplied by the identity matrix, which is then used as the covariance for a multivariate normal proposal density in the Metropolis Algorithm}

\item{priortheta}{only currently only option is "Exp", "Higs" and "none" will also be implemented}
}
\value{
returns a list containing mcmc.ma (samples) and accept (acceptance rates)
}
\description{
#### MCMC 3, multivariate normal distributed steps, each step is one p-dimensional vector from a multivariate normal distribution
}
\examples{
nsamp <- 100
burn <- 200
thin <- 10

n <- 10
x1 <- seq(-5,10,length.out = n)
x2 <- seq(0,15,length.out = n)
x <- expand.grid(x1,x2)
x <- as.matrix(x)
d2 <- c(0.01,0.2,0,0) #here we set the theta parameters to be 0.01 and 0.2.
# These are the modes of the distribution that we will sample from using MCMC
cor.par <- data.frame(matrix(data = d2,nrow = dim(x)[2],ncol = 2))
names(cor.par) <- c("Theta.y","Alpha.y")

R <- cor.matrix(x,cor.par) # obtain covariance matrix
L <- chol(R)
z <- as.matrix(rnorm(n^2))
y <- L\%*\%z

gp <- bceMCMC_mvrnorm(1000,10,10,x,y,reg = "constant",step =1, priortheta = "Exp")
mean(gp$mcmc.ma[,2]) #these means should be similar to the theta parameters set above
mean(gp$mcmc.ma[,1])
}

