% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor.matrix}
\alias{cor.matrix}
\title{Correlation Matrix for Gaussian Process based on Power-Exponential Correlation function
(c) Copyright William J. Welch 2000-2006.}
\usage{
cor.matrix(x, cor.par)
}
\arguments{
\item{x}{matrix/vector of explanatory variables}

\item{cor.par}{matrix of correlation parameters}
}
\value{
returns the correlation matrix based on power-exponential correlation structure
}
\description{
Correlation Matrix for Gaussian Process based on Power-Exponential Correlation function
(c) Copyright William J. Welch 2000-2006.
}
\examples{
n <- 5
x1 <- seq(-5,10,length.out = n)
x2 <- seq(0,15,length.out = n)

data1 <- expand.grid(x1,x2)
x <- data1
# create hyperparameter matrix of thetas and alphas, alphas set to 0 indicated guassian correlation
d2 <- c(0.01,0.2,0,0)
cor.par <- data.frame(matrix(data = d2,nrow = dim(x)[2],ncol = 2))
names(cor.par) <- c("Theta.y","Alpha.y")


R <- cor.matrix(data1,cor.par) # obtain covariance matrix

}

