% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_MCMC.R
\name{gpMCMC}
\alias{gpMCMC}
\title{master MCMC function}
\usage{
gpMCMC(nsamp, burn, thin, x, y, reg = "constant", step = 1,
  priortheta = "Exp", method = "uniform", plot = FALSE)
}
\arguments{
\item{nsamp}{number of MCMC samples to be generated}

\item{burn}{number of samples to be burned, must be less than nsamp}

\item{thin}{only include one of every 'thin' samples.  If thin is 10, only include one in every 10 samples}

\item{x}{covariate matrix/ vector x}

\item{y}{response vector}

\item{reg}{regression model, defaults to "constant"}

\item{step}{step length for MCMC, defaults to 1}

\item{priortheta}{prior, defaults to Exponential prior}

\item{method}{proposal distribution, defaults to "uniform"}

\item{plot}{defaults to FALSE}
}
\value{
returns a list containing the mcmc samples (mcmc_samples) and the acceptance rates for each covariate (acceptance)
}
\description{
master MCMC function
}

