% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmtx.R
\name{read.mtx}
\alias{read.mtx}
\title{For reading in matrix files generated by gasp Software}
\usage{
read.mtx(filename, sep = "")
}
\arguments{
\item{filename}{string, name of the file}

\item{sep}{by default set to nothing, but can also acommodate comma-separated values, or other}
}
\value{
returns the matrix that is read in
}
\description{
For reading in matrix files generated by gasp Software
}
\examples{
x <- c(1:3)
write.mtx(x,"x.mtx")
#x1 <- read.mtx("x.mtx")
#x1 == x
}

