% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writemtx.R
\name{write.mtx}
\alias{write.mtx}
\title{For writing out matrix files to be read by gasp software}
\usage{
write.mtx(x, filename, row.name = dimnames(x)[[1]],
  col.name = dimnames(x)[[2]], page.width = 80)
}
\arguments{
\item{x}{matrix to be written}

\item{filename}{intended name of matrix file, should end with ".mtx"}

\item{row.name}{optional vector indicating row-names, by default set to row-number}

\item{col.name}{optional vector indicating column-names, by default set to x1, x2, etc}

\item{page.width}{default = 80 characters}
}
\value{
returns nothing
}
\description{
For writing out matrix files to be read by gasp software
}
\examples{

x <- c(1:3)
write.mtx(x,"x.mtx")
}

